feat(noticias-eventos): add consolidated news and events data structure
- Introduced a new index file to manage news and events, allowing for easy addition and retrieval.
- Added individual news and events files for "Liderazgo Ambateño Junta 2026", "Sesión Solemne 2025", "Transmisión de Mando 2025", "Workshop Mes Juventud", and "Convención Nacional 2025".
- Implemented utility functions for filtering and counting news and events.

feat(proyectos): create project data structure and individual project files

- Established a new index file for managing projects, facilitating the addition of new projects.
- Created individual project files for "Conquistando Audiencias", "Construyendo Cultura Juntos", "Dale Otro Chance", "Feria del Libro", and "Líder a Líder".
- Added utility functions for filtering projects by area, state, and retrieving specific projects.

chore: add featured projects data

- Introduced a new file for featured projects, including details and utility functions for retrieval.
