+++ src/features/home/data/footer.ts
