+++ src/features/noticias-eventos/data/convencion-nacional-2025.ts
