+++ src/features/noticias-eventos/data/sesion-solemne-2025.ts
