+++ src/features/noticias-eventos/data/transmision-mando-2025.ts
