+++ src/features/proyectos/components/proyectos-por-area.tsx
@@ -3,11 +3,11 @@
 import Image from "next/image";
 import Link from "next/link";
 import { useState } from "react";
-import { getAllAreas } from "@/data/areas-oportunidad";
-import { getAllProyectos, getProyectosPorArea, getConteoProyectosPorArea } from "@/data/proyectos";
+import { getAllAreas } from "@/features/home/data/areas-oportunidad";
 import { Proyecto } from "@/types/proyecto";
 import { getAreaBySlug } from "@/types/enums";
 import { Button } from "@/components/ui/button";
+import { getAllProyectos, getConteoProyectosPorArea, getProyectosPorArea } from "@/features/proyectos/data";
 
 export default function ProyectosPorArea() {
   // Estado para controlar el botón activo (0 = Todos, 1 = Negocios, etc.)
